requirements.txt
astroid==2.0.4
autopep8==1.4.4
boto==2.49.0
boto3==1.9.159
botocore==1.12.159
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2
django-admin==1.3.2
django-allauth==0.38.0
django-cors-headers==2.4.0
django-excel-response2==2.0.8
django-filter==2.0.0
django-heroku==0.3.1
django-rest-auth==0.9.3
django-rest-passwordreset==1.0.0
django-six==1.0.4
django-storages==1.7.1
djangorestframework==3.9.0
djangotoolbox==1.8.0
docutils==0.14
gunicorn==19.9.0
idna==2.8
isort==4.3.4
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==2.1.0
Pillow==6.0.0
psycopg2==2.8.2
pycodestyle==2.5.0
Pygments==2.3.1
pylint==2.1.1
pylint-django==2.0.6
pylint-plugin-utils==0.5
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
s3transfer==0.2.0
screen==1.0.1
six==1.11.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.10.11
xlwt==1.3.0
